refactor: inline CGImage loading logic to BundleTextureImageLoader
